[fix] Termination check
If no termination criterion is specified, it defaults to
`MultiObjectiveDefaultTermination` which relies on the current Pareto
population. The problem is that on the first generation, `ARNSGA2` needs to
check termination before the Pareto population is set for the first time.

Now, termination is only checked from generation 2 onwards. This may lead to
transgressions when e.g. the maximum number of evaluations is lower than the
population size (but don't do that because it's stupid).