Resampled-ground performance indicators
New performance indicators `rggd`, `rggd+`, `rghv`, `rgigd`, `rgigd+`, which
are similar to resp. `ggd`, `ggd+`, `ghv`, `gigd`, `gigd+` except that the
ground problem is resampled a given number of times before calculation (see
`rg_n_eval` in the problem description dicts).

Technically speaking, `rggd` with `rg_n_eval = 1` is *not* the same as simply
`ggd` because the ground problem is reevaluated regardless. This is to prevent
undue correlation between `rggd` and `ggd` if both are used.